Move GetStudentProfiles to routines
